+++ code/programs/haskell/build-tool/test/Spec.hs
@@ -2,7 +2,7 @@ import Test.Hspec
 
 import BuildToolSpec (buildToolSpec)
 import HashingSpec (hashingSpec)
-import ResolutionUtf8Spec (resolutionCabalSpec, resolutionPythonSpec, resolutionRubySpec, resolutionRustSpec, resolutionUtf8Spec)
+import ResolutionUtf8Spec (resolutionCabalSpec, resolutionPerlSpec, resolutionPythonSpec, resolutionRubySpec, resolutionRustSpec, resolutionUtf8Spec)
 
 main :: IO ()
 main = hspec spec
@@ -16,3 +16,4 @@ spec = do
     resolutionPythonSpec
     resolutionRustSpec
     resolutionRubySpec
+    resolutionPerlSpec